+++ backend/src/main/java/blok2/http/controllers/TimeslotController.java
@@ -1,9 +1,9 @@
 package blok2.http.controllers;
 
 import blok2.database.dao.ITimeslotDao;
-import blok2.http.security.authorization.AuthorizedLocationController;
 import blok2.exceptions.InvalidRequestParametersException;
 import blok2.exceptions.NoSuchDatabaseObjectException;
+import blok2.http.controllers.authorization.AuthorizedLocationController;
 import blok2.model.calendar.Timeslot;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -13,8 +13,6 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.time.LocalDate;
 import java.util.List;
-
-import static org.springframework.http.HttpStatus.BAD_REQUEST;
 
 @RestController
 @RequestMapping("locations/timeslots")
@@ -31,8 +29,10 @@ public class TimeslotController extends AuthorizedLocationController {
     @PreAuthorize("permitAll()")
     public Timeslot getTimeslot(@PathVariable("timeslotId") int timeslotId) {
         Timeslot timeslot = timeslotDAO.getTimeslot(timeslotId);
-        if (timeslot == null)
+
+        if (timeslot == null){
             throw new NoSuchDatabaseObjectException("Timeslot does not exist");
+        }
 
         return timeslot;
     }
@@ -41,7 +41,8 @@ public class TimeslotController extends AuthorizedLocationController {
     @PreAuthorize("permitAll()")
     public List<Timeslot> getTimeslotsOfLocation(@PathVariable("locationId") int locationId) {
         // Only return timeslots from up to 6 months old until all in the future, to limit data sent to the frontend.
-        return timeslotDAO.getTimeslotsOfLocationAfterTimeslotDate(locationId, LocalDate.now().minusMonths(6));
+        LocalDate sixMonthsAgo = LocalDate.now().minusMonths(6);
+        return timeslotDAO.getTimeslotsOfLocationAfterTimeslotDate(locationId, sixMonthsAgo);
     }
 
     @PutMapping()
@@ -49,16 +50,18 @@ public class TimeslotController extends AuthorizedLocationController {
     public ResponseEntity<Object> updateTimeslot(@Valid @RequestBody Timeslot timeslot) {
         Timeslot original = timeslotDAO.getTimeslot(timeslot.getTimeslotSeqnr());
         // Validate original location
-        isAuthorized(original.getLocationId());
+        this.checkLocationAuthorization(original.getLocationId());
         // Validate future location
-        isAuthorized(timeslot.getLocationId());
+        this.checkLocationAuthorization(timeslot.getLocationId());
 
         // Check that new number of seats is not less than number of reservations
         if (timeslot.getSeatCount() < original.getAmountOfReservations()) {
-            return new ResponseEntity<>(new InvalidRequestParametersException("Number of seats cannot be less than number of reservations"), BAD_REQUEST);
+            throw new InvalidRequestParametersException(
+                "Number of seats cannot be less than number of reservations"
+            );
         }
 
-        // Execute update
+        // Update the timeslot.
         timeslotDAO.updateTimeslot(timeslot);
 
         return ResponseEntity.ok(timeslot);
@@ -68,7 +71,7 @@ public class TimeslotController extends AuthorizedLocationController {
     @DeleteMapping
     @PreAuthorize("hasAuthority('HAS_AUTHORITIES') or hasAuthority('ADMIN')")
     public void deleteTimeslot(@RequestBody Timeslot timeslot) {
-        isAuthorized(timeslot.getLocationId());
+        checkLocationAuthorization(timeslot.getLocationId());
 
         timeslotDAO.deleteTimeslot(timeslot);
     }
@@ -76,7 +79,7 @@ public class TimeslotController extends AuthorizedLocationController {
     @PostMapping
     @PreAuthorize("hasAuthority('HAS_AUTHORITIES') or hasAuthority('ADMIN')")
     public Timeslot addTimeslot(@RequestBody Timeslot timeslot) {
-        isAuthorized(timeslot.getLocationId());
+        checkLocationAuthorization(timeslot.getLocationId());
 
         return timeslotDAO.addTimeslot(timeslot);
     }
@@ -85,7 +88,7 @@ public class TimeslotController extends AuthorizedLocationController {
     @PreAuthorize("hasAuthority('HAS_AUTHORITIES') or hasAuthority('ADMIN')")
     public Timeslot setRepeatable(@PathVariable("timeslotId") Integer timeslotSequenceNumber, @RequestBody SetRepeatableBodyDao rep) {
         Timeslot ts = timeslotDAO.getTimeslot(timeslotSequenceNumber);
-        isAuthorized(ts.getLocationId());
+        checkLocationAuthorization(ts.getLocationId());
         ts.setRepeatable(rep.repeatable);
         return timeslotDAO.updateTimeslot(ts);
     }
